requirements.txt
avro==1.10.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.2
chardet==3.0.4
charset-normalizer==2.0.6
click==8.1.3
cloudevents==1.7.1
cryptography==3.1
decorator==4.4.2
deprecation==2.1.0
fastavro==1.3.3
Flask==2.2.2
functions-framework==3.2.1
google-api-core==1.33.2
google-api-python-client==2.66.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
google-cloud-appengine-logging==1.1.6
google-cloud-audit-log==0.2.4
google-cloud-bigquery==2.34.4
google-cloud-bigquery-storage==2.3.0
google-cloud-core==1.4.1
google-cloud-logging==2.7.2
google-cloud-spanner==3.10.0
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.57.0
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.48.2
gunicorn==20.1.0
httplib2==0.18.1
idna==3.2
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
libcst==0.3.13
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.19.0
oauthlib==3.1.0
packaging==21.0
pandas==1.0.5
proto-plus==1.22.1
protobuf==3.20.3
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==12.8
pytz==2020.1
PyYAML==5.3.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tornado==6.0.4
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.7
watchdog==2.1.9
Werkzeug==2.2.2
zipp==3.11.0
